kits: add ugv devkit v1 chassis bringup
