+++ src/kits/README.md
@@ -1,0 +1,18 @@
+# Overview
+This folder contains "bringup" packages for the different sensor/chassis kits we support for the [Weston Robot Development Kit](https://docs.westonrobot.net/wr_dev_kit/wr_dev_kit.html) (devkit). 
+
+## Kits
+* [Chassis](./chassis/)
+  * Devkit chassis and frame
+* [Top Sensor Kits](./top/)
+  * Sensor kits designed to be mounted on the top of the devkit chassis
+* [Base Sensor Kits](./base/)
+  * Sensor kits designed to be mounted on the robot base itself.
+
+## Supported Kits
+* Chassis Kits
+  * [UGV Devkit V1.0](https://docs.westonrobot.net/wr_dev_kit/ugv_dev_kit_v1.html)
+* Top Sensor Kits
+  * [Mid360 Lidar + IMU](./top/mid360_sensor_kit_bringup/)
+* Base Sensor Kits
+  * [W200d ultrasonic](./base/w200d_sensor_kit_bringup/)
