+++ src/kits/base/w200d_sensor_kit_bringup/package.xml
@@ -12,6 +12,10 @@
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
 
+  <exec_depend>wrp_ros2</exec_depend>
+  <exec_depend>xacro</exec_depend>
+  <exec_depend>robot_state_publisher</exec_depend>
+
   <export>
     <build_type>ament_cmake</build_type>
   </export>
