+++ src/kits/chassis/ugv_devkit_v1_bringup/meshes/wr_devkit_chassis_link.stl
