+++ src/kits/top/mid360_sensor_kit_bringup/config/MID360_config.json
