+++ src/kits/top/mid360_sensor_kit_bringup/meshes/mid360_sensor_kit_base_link.stl
