+++ src/kits/top/mid360_sensor_kit_bringup/package.xml
@@ -15,6 +15,7 @@
   <exec_depend>wrp_ros2</exec_depend>
   <exec_depend>livox_ros_driver2</exec_depend>
   <exec_depend>xacro</exec_depend>
+  <exec_depend>robot_state_publisher</exec_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
