--- src/wr_devkit_description/urdf/materials.xacro
@@ -1,8 +1,0 @@
-<?xml version="1.0" ?>
-<robot name="wr_devkit" xmlns:xacro="http://www.ros.org/wiki/xacro" >
-
-<material name="silver">
-  <color rgba="0.700 0.700 0.700 1.000"/>
-</material>
-
-</robot>
