--- src/wr_devkit_description/urdf/mid360_sensorkit.urdf.xacro
@@ -1,47 +1,0 @@
-<?xml version="1.0"?>
-<robot name="mid360_sensorkit" xmlns:xacro="http://www.ros.org/wiki/xacro">
-
-  <link name="mid360_sensorkit_base_link">
-    <inertial>
-      <origin xyz="0.0 0.0 0.0" rpy="0 0 0" />
-      <mass value="0.0" />
-      <inertia ixx="0.0" iyy="0.0" izz="0.0" ixy="0.0" iyz="0.0" ixz="0.0" />
-    </inertial>
-    <visual>
-      <origin xyz="0 0 0" rpy="0 0 0" />
-      <geometry>
-        <mesh
-          filename="package://wr_devkit_description/meshes/mid360_sensorkit_base_link.stl"
-          scale="0.001 0.001 0.001" />
-      </geometry>
-      <material name="silver" />
-    </visual>
-    <collision>
-      <origin xyz="0 0 0" rpy="0 0 0" />
-      <geometry>
-        <mesh
-          filename="package://wr_devkit_description/meshes/mid360_sensorkit_base_link.stl"
-          scale="0.001 0.001 0.001" />
-      </geometry>
-    </collision>
-  </link>
-
-  <link name="ch104m_imu_frame">
-  </link>
-
-  <link name="livox_frame">
-  </link>
-
-  <joint name="ch104m_imu_joint" type="fixed">
-    <origin xyz="0.0055 0.0 0.087483" rpy="0 0 1.5708" />
-    <parent link="mid360_sensorkit_base_link" />
-    <child link="ch104m_imu_frame" />
-  </joint>
-
-  <joint name="livox_joint" type="fixed">
-    <origin xyz="0.0055 0.0 0.087483" rpy="0 0 0" />
-    <parent link="mid360_sensorkit_base_link" />
-    <child link="livox_frame" />
-  </joint>
-
-</robot>
